fix(plugin-async-task-manager): add try catch for cleanup (#8215)
* fix(plugin-async-task-manager): add try catch for cleanup

* fix(plugin-async-task-manager): fix try / catch

+++ packages/plugins/@nocobase/plugin-async-task-manager/src/server/base-task-manager.ts
@@ -135,34 +135,47 @@ export class BaseTaskManager implements AsyncTasksManager {
   private cleanup = async () => {
     this.logger.debug('Running cleanup for completed tasks...');
     const TaskRepo = this.app.db.getRepository('asyncTasks');
-    const tasksToCleanup = await TaskRepo.find({
-      fields: ['id'],
-      filter: {
-        $or: [
-          {
-            status: [TASK_STATUS.SUCCEEDED, TASK_STATUS.FAILED],
-            doneAt: {
-              $lt: new Date(Date.now() - this.cleanupDelay),
+    try {
+      const tasksToCleanup = await TaskRepo.find({
+        fields: ['id'],
+        filter: {
+          $or: [
+            {
+              status: [TASK_STATUS.SUCCEEDED, TASK_STATUS.FAILED],
+              doneAt: {
+                $lt: new Date(Date.now() - this.cleanupDelay),
+              },
             },
-          },
-          {
-            status: TASK_STATUS.CANCELED,
-          },
-        ],
-      },
-    });
+            {
+              status: TASK_STATUS.CANCELED,
+            },
+          ],
+        },
+      });
 
-    this.logger.debug(`Found ${tasksToCleanup.length} tasks to cleanup`);
+      this.logger.debug(`Found ${tasksToCleanup.length} tasks to cleanup`);
 
-    for (const task of tasksToCleanup) {
-      this.tasks.delete(task.id);
-      this.progressThrottles.delete(task.id);
+      if (tasksToCleanup.length) {
+        for (const task of tasksToCleanup) {
+          this.tasks.delete(task.id);
+          this.progressThrottles.delete(task.id);
+        }
+
+        await TaskRepo.destroy({
+          filterByTk: tasksToCleanup.map((task) => task.id),
+          individualHooks: true,
+        });
+      }
+    } catch (error) {
+      this.logger.error(`DB error during cleanup: ${error.message}. Will stop cleanup timer.`, { error });
+
+      if (error.name === 'SequelizeConnectionAcquireTimeoutError') {
+        if (this.cleanupTimer) {
+          clearInterval(this.cleanupTimer);
+          this.cleanupTimer = null;
+        }
+      }
     }
-
-    await TaskRepo.destroy({
-      filterByTk: tasksToCleanup.map((task) => task.id),
-      individualHooks: true,
-    });
   };
 
   private getThrottledProgressEmitter(taskId: string, userId: number) {
